Use mutex to guard access to the client buffers
