+++ pkg/client/client.go
@@ -50,9 +50,10 @@ type DataSetClient struct {
 	Config           *config.DataSetConfig
 	Client           *http.Client
 	SessionInfo      *add_events.SessionInfo
-	buffer           sync.Map
+	buffer           map[string]*buffer.Buffer
 	buffersEnqueued  atomic.Uint64
 	buffersProcessed atomic.Uint64
+	buffersMutex     sync.Mutex
 	PubSub           *pubsub.PubSub
 	workers          sync.WaitGroup
 	LastHttpStatus   atomic.Uint32
@@ -98,8 +99,10 @@ func NewClient(cfg *config.DataSetConfig, client *http.Client, logger *zap.Logge
 		Id:               id,
 		Config:           cfg,
 		Client:           client,
+		buffer:           make(map[string]*buffer.Buffer),
 		buffersEnqueued:  atomic.Uint64{},
 		buffersProcessed: atomic.Uint64{},
+		buffersMutex:     sync.Mutex{},
 		PubSub:           pubsub.New(0),
 		workers:          sync.WaitGroup{},
 		LastHttpStatus:   atomic.Uint32{},
@@ -137,21 +140,17 @@ func NewClient(cfg *config.DataSetConfig, client *http.Client, logger *zap.Logge
 func (client *DataSetClient) Buffer(key string, info *add_events.SessionInfo) *buffer.Buffer {
 	session := fmt.Sprintf("%s-%s", client.Id, key)
 	// get buf so we can start using it
-	buf, loaded := client.buffer.LoadOrStore(
-		session,
-		buffer.NewEmptyBuffer(session, client.Config.Tokens.WriteLog),
-	)
-
-	buff := buf.(*buffer.Buffer)
-
-	if !loaded {
-		client.initBuffer(buff, info)
-
-		// it's brand-new buf, so let's create subscriber
+	client.buffersMutex.Lock()
+	defer client.buffersMutex.Unlock()
+	buf, found := client.buffer[session]
+	if !found {
+		buf = buffer.NewEmptyBuffer(session, client.Config.Tokens.WriteLog)
+		client.initBuffer(buf, info)
+		client.buffer[session] = buf
 		client.AddEventsSubscriber(session)
 	}
 
-	return buff
+	return buf
 }
 
 func (client *DataSetClient) initBuffer(buff *buffer.Buffer, info *add_events.SessionInfo) {
@@ -258,7 +257,7 @@ func (client *DataSetClient) ListenAndSendBufferForSession(session string, ch ch
 					}
 
 					client.sleep(retryAfter, buf)
-					buf.Status.Store(uint32(buffer.Retrying))
+					buf.SetStatus(buffer.Retrying)
 					buf.Attempt++
 
 					// and publish message back
@@ -279,23 +278,22 @@ func (client *DataSetClient) ListenAndSendBufferForSession(session string, ch ch
 func (client *DataSetClient) bufferSweeper(delay time.Duration) {
 	client.Logger.Info("Starting buffer sweeper with delay", zap.Duration("delay", delay))
 	for i := uint64(0); ; i++ {
+		if client.finished.Load() {
+			client.Logger.Info("Stopping buffer sweeper", zap.Uint64("sweepId", i))
+			break
+		}
 		kept := atomic.Uint64{}
 		swept := atomic.Uint64{}
 		client.Logger.Debug("Buffer sweeping started", zap.Uint64("sweepId", i))
-		client.buffer.Range(func(k, v interface{}) bool {
-			buf, ok := v.(*buffer.Buffer)
-			if ok {
-				if buf.ShouldSendAge(delay) {
-					client.PublishBuffer(buf)
-					swept.Add(1)
-				} else {
-					kept.Add(1)
-				}
+		buffers := client.buffer
+		for _, buf := range buffers {
+			if buf.HasStatus(buffer.Ready) && buf.ShouldSendAge(delay) {
+				client.PublishBuffer(buf)
+				swept.Add(1)
 			} else {
-				client.Logger.Error("Unable to convert message to buffer")
+				kept.Add(1)
 			}
-			return true
-		})
+		}
 
 		// log just every n-th sweep
 		lvl := zap.DebugLevel
@@ -316,9 +314,11 @@ func (client *DataSetClient) bufferSweeper(delay time.Duration) {
 }
 
 func (client *DataSetClient) PublishBuffer(buf *buffer.Buffer) {
-	originalStatus := buffer.Status(buf.Status.Load())
-	buf.Status.Store(uint32(buffer.Publishing))
-	if originalStatus == buffer.Ready {
+	client.buffersMutex.Lock()
+	defer client.buffersMutex.Unlock()
+	originalStatus := buf.Status()
+	buf.SetStatus(buffer.Publishing)
+	if originalStatus == buffer.Ready || originalStatus == buffer.AddingBundles {
 		newBuf := buffer.NewEmptyBuffer(buf.Session, buf.Token)
 		client.Logger.Debug(
 			"Removing buffer for session",
@@ -328,7 +328,7 @@ func (client *DataSetClient) PublishBuffer(buf *buffer.Buffer) {
 		)
 
 		client.initBuffer(newBuf, buf.SessionInfo())
-		client.buffer.Store(buf.Session, newBuf)
+		client.buffer[buf.Session] = newBuf
 	}
 	client.Logger.Debug("publishing buffer", buf.ZapStats()...)
